+++ tests/test_basic.py
@@ -10,7 +10,7 @@ from time import sleep
 import fsspec
 import yaml
 
-from yamlinclude import YamlIncludeCtor
+from yaml_include import Constructor
 
 from ._internal import YAML1, YAML2, YAML_LOADERS
 
@@ -157,7 +157,7 @@ class DefaultFsBasicTestCase(BaseTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        ctor = YamlIncludeCtor(base_dir="tests/data")
+        ctor = Constructor(base_dir="tests/data")
         for loader_cls in YAML_LOADERS:
             yaml.add_constructor("!inc", ctor, loader_cls)
 
@@ -181,7 +181,7 @@ class FileFsBasicTestCase(BaseTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        ctor = YamlIncludeCtor(fs=fsspec.filesystem("file"), base_dir=lambda: "tests/data")
+        ctor = Constructor(fs=fsspec.filesystem("file"), base_dir=lambda: "tests/data")
         for loader_cls in YAML_LOADERS:
             yaml.add_constructor("!inc", ctor, loader_cls)
 
@@ -248,7 +248,7 @@ class SimpleHttpBasicTestCase(BaseTestCase):
         cls.server_thread.start()
         sleep(1)
         host, port = httpd.socket.getsockname()[:2]
-        ctor = YamlIncludeCtor(
+        ctor = Constructor(
             fs=fsspec.filesystem("http", client_kwargs=dict(base_url=f"http://{host}:{port}")),
             base_dir="/tests/data",
         )
@@ -291,7 +291,7 @@ class DefaultFsNoBaseDirBasicTestCase(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        ctor = YamlIncludeCtor()
+        ctor = Constructor()
         for loader_cls in YAML_LOADERS:
             yaml.add_constructor("!inc", ctor, loader_cls)