+++ tests/test_loadfunc.py
@@ -2,17 +2,17 @@ import unittest
 from textwrap import dedent
 
 import yaml
-from yamlinclude import (
-    YamlIncludeCtor,
-    yamlinclude_load,
-    yamlinclude_lazy_load,
+from yaml_include import (
+    Constructor,
+    load,
+    lazy_load,
 )
 
 from ._internal import YAML_LOADERS, YAML1, YAML2
 
 
 class LoadFuncTestCase(unittest.TestCase):
-    ctor = YamlIncludeCtor(base_dir="tests/data", autoload=False)
+    ctor = Constructor(base_dir="tests/data", autoload=False)
 
     @classmethod
     def setUpClass(cls):
@@ -37,7 +37,7 @@ class LoadFuncTestCase(unittest.TestCase):
         ).strip()
         for loader_cls in YAML_LOADERS:
             d0 = yaml.load(yaml_string, loader_cls)
-            d1 = yamlinclude_load(d0, loader_cls, self.ctor)
+            d1 = load(d0, loader_cls, self.ctor)
             self.assertDictEqual(YAML1, d1["list"][0])
             self.assertDictEqual(YAML2, d1["list"][1])
             self.assertDictEqual(YAML1, d1["dict"]["yaml1"])
@@ -56,7 +56,7 @@ class LoadFuncTestCase(unittest.TestCase):
         ).strip()
         for loader_cls in YAML_LOADERS:
             d0 = yaml.load(yaml_string, loader_cls)
-            yamlinclude_load(d0, loader_cls, self.ctor, inplace=True)
+            load(d0, loader_cls, self.ctor, inplace=True)
             self.assertDictEqual(YAML1, d0["list"][0])
             self.assertDictEqual(YAML2, d0["list"][1])
             self.assertDictEqual(YAML1, d0["dict"]["yaml1"])
@@ -70,7 +70,7 @@ class LoadFuncTestCase(unittest.TestCase):
         ).strip()
         for loader_cls in YAML_LOADERS:
             d0 = yaml.load(yaml_string, loader_cls)
-            d1 = yamlinclude_load(d0, loader_cls, self.ctor, nested=True)
+            d1 = load(d0, loader_cls, self.ctor, nested=True)
             self.assertDictEqual(YAML1, d1["nested"]["list"][0])
             self.assertDictEqual(YAML2, d1["nested"]["list"][1])
             self.assertDictEqual(YAML1, d1["nested"]["dict"]["yaml1"])
@@ -84,15 +84,15 @@ class LoadFuncTestCase(unittest.TestCase):
         ).strip()
         for loader_cls in YAML_LOADERS:
             d0 = yaml.load(yaml_string, loader_cls)
-            yamlinclude_load(d0, loader_cls, self.ctor, inplace=True, nested=True)
+            load(d0, loader_cls, self.ctor, inplace=True, nested=True)
             self.assertDictEqual(YAML1, d0["nested"]["list"][0])
             self.assertDictEqual(YAML2, d0["nested"]["list"][1])
             self.assertDictEqual(YAML1, d0["nested"]["dict"]["yaml1"])
             self.assertDictEqual(YAML2, d0["nested"]["dict"]["yaml2"])
 
 
-class IterableLoadFuncTestCase(unittest.TestCase):
-    ctor = YamlIncludeCtor(base_dir="tests/data", autoload=False)
+class LazyLoadFuncTestCase(unittest.TestCase):
+    ctor = Constructor(base_dir="tests/data", autoload=False)
 
     @classmethod
     def setUpClass(cls):
@@ -117,7 +117,7 @@ class IterableLoadFuncTestCase(unittest.TestCase):
         ).strip()
         for loader_cls in YAML_LOADERS:
             d0 = yaml.load(yaml_string, loader_cls)
-            for _ in yamlinclude_lazy_load(d0, loader_cls, self.ctor):
+            for _ in lazy_load(d0, loader_cls, self.ctor):
                 pass
             self.assertDictEqual(YAML1, d0["list"][0])
             self.assertDictEqual(YAML2, d0["list"][1])
@@ -132,7 +132,7 @@ class IterableLoadFuncTestCase(unittest.TestCase):
         ).strip()
         for loader_cls in YAML_LOADERS:
             d0 = yaml.load(yaml_string, loader_cls)
-            for _ in yamlinclude_lazy_load(d0, loader_cls, self.ctor, nested=True):
+            for _ in lazy_load(d0, loader_cls, self.ctor, nested=True):
                 pass
             self.assertDictEqual(YAML1, d0["nested"]["list"][0])
             self.assertDictEqual(YAML2, d0["nested"]["list"][1])